Table HTTP status added part 2

+++ backend/src/main/java/com/pants/backend/controller/TableController.java
@@ -14,6 +14,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 
 import java.util.List;
+import java.util.Map;
+
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.PutMapping;
@@ -73,21 +75,30 @@ public class TableController {
 
         if (existingTable == null) {
             return ResponseEntity
-            .status(HttpStatus.NOT_FOUND)
-            .body(new ErrorResponse(404, "Table not found"));
+                    .status(HttpStatus.NOT_FOUND)
+                    .body(new ErrorResponse(404, "Table not found"));
         }
 
         existingTable.setTableNumber(table.getTableNumber());
         existingTable.setCapacity(table.getCapacity());
 
-        Table updatedTable =tableRepository.save(existingTable);
+        Table updatedTable = tableRepository.save(existingTable);
 
         return ResponseEntity.ok(updatedTable);
     }
 
     @DeleteMapping("/{id}")
-    public void deleteTable(@PathVariable Integer id) {
-        tableRepository.deleteById(id);
+    public ResponseEntity<?> deleteTable(@PathVariable Integer id) {
+
+        if (tableRepository.existsById(id)) {
+            tableRepository.deleteById(id);
+
+            return ResponseEntity.ok(Map.of("message", "Successfully deleted table with id " + id));
+        }
+
+        return ResponseEntity
+                .status(HttpStatus.NOT_FOUND)
+                .body(new ErrorResponse(404, "Table not found"));
     }
 
 }
